requirements.txt
soda-core-bigquery==3.0.32
google-cloud-bigquery-storage==2.19.1
pyarrow==11.0.0